Extract static class ResolveChangeDamage
